Restructure `ElementType` and `Document`
This stems redundant constructs, such as `Root`, `Body`, and `Div` being
merged into `Container`, as the exact structure is unnecessary.

This might mean that CSS selectors should be applied before this stage,
but that's a problem for a later time.

`Bold` is also merged with `Text` as they serve the same purpose, bold
should be represented as a style, not as an element type.

Furthermore, `Input`, `Button`, `Image`, `Link`, `List`, and `Table` are
all potential candidates as HTML element types.

With `Component` also being suggested, but I am unsure how it would fit
as a distinct element type.

---

The goal of an element type is to represent the type of HTML tags,
however, some loss of information is desired, being represented through
meta-data about the specific item; for example, <b> could be represented
as a `ElementType::Text` with a bold styling to it.