Document::parse: Return an error when parsing text in a non-text element
This creates a new `DOMParseError` type `UnexpectedSymbol` that is
returned when a non-control character is found in an element with a type
not equal to `ElementType::Text`.

Which basically means a (probably) text character in (probably) an
`ElementType::Container`.

This previously was ignored, and then (probably) given an implicit text
child element, but who knows—it was undefined behavior.

---

This commit also modifies a test case that expects that improper
behavior.