Fold whitespace in tag attributes
This creates a function `skip_spaces` that skips through an iterator till
it finds a non-whitespace character then it returns.

The tag parser calls the function at each occurance of whitespace past
the first one, the first whitespace remarks the end of the tag type. Then
the following text is for the attributes.

Creating unit tests for the function is necessary. It should also be used
in parsing text content.