Converts `tokenize::tokenize` into an Iterator and implement some states
This implements:
- State::Data;
- State::RcData, and;
- State::TagOpen.

The implementations are incomplete, and were only trial implementations
to understand the proper form the parse should be in; I believe it is
in an acceptable form that is terse and capable enough.

Some rationale behind the choices includes:

- The outer API (`Lexer::parse`) is a method of access to the iterator
tokenizer to allow for the tokenizer to execute as data is being read,
this is certainly not done with enough clarity that it could be rational
but it is an attempt..
- The Iterator returning Option<()> is because there is nothing to
return, tokens are appended directly because often times we must append
multiple tokens, and the same goes for errors.
- The states are one large match statement because it is too early to
draw patterns within (for me), however, once patterns emerge I will
aggregate behavior. However, I see that having it as one large match
statemenet is desirable, as it will allow for better readability;
although the match statement will be long, the relevant arm is
particularly small, and there is little need to complicate it, but what
do I know.