Fix bug on not consuming a character
This happens because we delegated consuming a character within the
specific character handling, when it is usually handled when handling the
state; this discrepancy is because certain characters require
reconsumption in the upcoming state.

This error is frequent, and requires making reconsumption explicit,
rather than having consumption be specified in the branch. Which can not
be done when using a `Iterator<Item = char>`.

Eventually, the character iterator should be replaced with something more
like a `Vec<char>` to be able to index it in constant time.